Fix vinyl/coalesce test hang
Slices make range size estimation inaccurate. This may lead to the
coalesce test hang waiting for all ranges to be coalesced in case the
estimated sum range size turns out to be greater than the threshold
(range_size / 2). To make sure it never happens, let's make the test
more aggressive: let it not only delete 90% keys, but also strip the
rest of padding.