vinyl: fix scheduler hang if quota is exceeded but there's nothing to dump
The quota may be exceeded even if all in-memory trees are empty, because
quota is used by pending transactions. If this happens, the scheduler
will keep dumping empty trees in a tight loop w/o yielding, thus not
even letting pending transactions to fail on timeout. Fix this by adding
a sanity check to vy_scheduler_peek_dump() assuring there's something we
can dump before proceeding to scheduling a task.