Add vinyl/upgrade test
The test checks that tarantool can start from data generated by older
versions. Currently, we only check against 1.7.4, because this is when
the vinyl format was frozen and it hasn't changed since then, but we
will add sub test cases every time we extend the format. The test is
shipped with a script (fill.lua) that generates test data.

Closes #2407