box: rework internal gc API
The current API only allows to get the vclock of the last checkpoint,
optionally referencing it (gc_last_checkpoint, gc_ref_last_checkpoint),
while in order to keep track of the checkpoint used by a replica we need
to be able to lookup a checkpoint by vclock. If referenced a checkpoint
must be eventually unreferenced with gc_unref_checkpoint, which takes a
vclock. This is confusing, because we actually deal not with vclocks,
but with checkpoint_info objects. That said, let's rework the internal
API as follows:

 - Make gc_last_checkpoint() return checkpoin_info object instead
   of vclock.
 - Add gc_lookup_checkpoint() to lookup a checkpoint_info object
   by vclock.
 - Replace gc_ref_last_checkpoint() and gc_unref_checkpoint()
   with checkpoint_ref() and checkpoint_unref(), both taking a
   checkpoint_info object.