Fix vinyl garbage collection tests
After the common memory level was introduced, compaction and dump can
proceed in parallel for the same range and snapshot doesn't block
compaction, as it used to. As a result, the following sequence of
commands

  s = box.schema.space.create('test', {engine = 'vinyl'})
  s:create_index('pk', {run_count_per_level = 1})
  s:insert{1} box.snapshot() -- dump
  s:insert{2} box.snapshot() -- dump + compaction
  box.internal.gc.run(box.info.signature)

can result in a different number of run files in the index directory (1
or 3) depending on whether compaction manages to finish before snapshot
completion or not. This breaks garbage collection tests. Fix this by
forcing a snapshot (via a memtx table) before any invocation of garbage
collection.