vinyl: fix use after free in initial join
vy_join_cb sets vy_join_ctx::key_def to point to vy_log_record::key_def.
This is incorrect, as vy_log_record is only guaranteed to be valid until
the callback returns. Using it like that leads to use-after-free in
vy_join() => vy_send_range(), because the recovery object is already
deleted there. Fix this by copying the key_def to the join context.