coio: merge coeio and coio
Originally, coio was used for socket io based on libev, and coeio
for everything based on libeio. This seems to be hard to hard to remember as
the previous commit demonstrates.

Some of the APIs are already mixed up and use the wrong prefix (see,
for example, struct coio_task).

Merge the two set of calls together and use the same prefix for all.

First patch in the set, renaming the files and adjusting header guards
and makefiles.