vinyl: fix broken rollback to savepoint
Now vinyl's rollback to TX savepoint doesn't work in case when one
TX has several statements modifying the same key.

A simple example is presented in gh ticket 2589.

Fix it.

Add a big test that compares vinyl with memtx in case of many
failing (catched with pcall) statements done in a transaction.

Fixes #2589