sql: make indexes covering, fix scan cost, reorder indexes list
This patch was implemented w/ @Gerold103

The decision to reorder indexes list was made due to several reasons:
 - indexes are sorted this way in Tarantool internal representation
 - if all indexes become covering, then Tarantool should try pk with
     higher priority, but it seems that in some cases sqlite chooses
     the first suitable index from the indexes list

Tarantool indexes should be covering by nature, because it's representation
is just a set pointers in to a tuple (containing all columns of the row)

This change led to reading data from secondary indexes on scan if possible.
To increase usage of locality of data it was necessary to
change cost calculation formula.
By now scan over pk is always better than scan over secondary index for
Tarantool table, stored in memory.

This change also affected some tests. All result changes was reviewed and
none of them was inefficient.
main changes:
 - "COVERING" word appeared in each query plan statement
   (because indexes became covering)
 - sqlite engine changed some decisions on which index to use for full scan
   (this affected to order of results in some queries)

Closes #2110