vinyl: do not force recovery of vylog
Swallowing errors while reading a vylog file is extremely dangerous,
because vylog files contain metadata layout. Any error in vylog might
result in a data corruption beyond recovery in the future. So we must
fail recovery/snapshot if we fail to read the current vylog file. If
this happens, data can still be restored by reading the previous vylog
file or by recovering vylog using write iterator (the latter is yet to
be implemented).