sql: Enable generic transactions
Emit 'OP_TTransaction' opcode while compiling 'BEGIN' SQL statement.
Also, do Tarantool's rollback in if 'OR ROLLBACK' modifier was
specified in SQL statement and it failed to complete.

This enables transactions with simple statements inside to work.
By simple it is meant that they're inserting only one row, or have no
triggers, since w/o in-stmt transactions, changes of single SQL
statement won't be discarded. This limitation is due to #2025
and should be resolved in current release cycle.

Also, fully uncomment and convert sql-tap/trigger9.test.lua

TODO: This patch temporary removes DDL from within transactions.
Need to revert this part when support of DDL in transactions
is checked in.

Closes #2140