cbus: fix racy cbus_endpoint->n_pipes decrement
cbus_endpoint->n_pipes is incremented from a thread connecting to the
endpoint (see cpipe_create()) while it is decremented from the thread
hosting the endpoint (see cbus_endpoint_poison_f()). While it is
protected with cbus->mutex in the former case, it is decremented
locklessly, which can result in a race leading to n_pipes never reaching
0 and hence cbus_endpoint_destroy() never returning. Fix this by taking
cbus->mutex for decrementing the variable.