vinyl: fix assertion in cache add
vinyl_cache_add takes pair of statements - current and previous.
By design, there must be no statements in cache between them.

Some time ago a check was made that the previous statement is in
cache and lays right next to current statement; if that was not
true, the function exited without building a cache chain.
The was wrong because there are some cases when the statement
could be removed by TX finalize/rollback after prepare or even by
cache garbage collector.

Later in this check 'return' was replaced by 'assert' and that
caused rare test crash.

Fix this check. Check only two things 1) absence of entries between
current and previous statements 2) presence of previous statement
if current is already in chain.

Add a test case.

See code comments for details.

Fixes #2685