vinyl: fix read iterator restoration to a newer version of the same key
After the read iterator selects the minimal key across all available
sources, it checks mutable sources for new statements using ->restore()
callback. If there is a new statement in a source, it uses it as the min
key provided it is *strictly* less than the current min key. If they are
equal, the min key isn't changed, but this is wrong, because the new
statement may be newer than the statement selected previously. If we
don't select it, we might end up with stale data in the cache. Fix this.