Store tuple format in struct space
Currently, the format is stored in Handler and can be retrieved with
Handler::format(). The only reason for such a layer of indirection is
that the format should be created by an Engine method and currently it
is created by Engine::createSpace(). Let's introduce a separate method
for create a space format, Engine::createFormat(), and move the format
back to where it belongs, i.e. to struct space.

This is a step towards converting class Handler to struct space_vtab.

Needed for #2776