alter: restrict space format altering
Disallow changing field type of space format fields to an
incompatible type.
For example, forbid changing a type from integer to string,
but allow to change from integer to number.

If an user adds new fields, then validate existing tuples to be
compatible with a new format. Check works in O(N) time and for
big spaces can be realy long. But it can be skipped, if conditions
below are ok:
1. if space format field count < old tuple format field count. Then
   field count of all tuples >= new format field count;
2. new fields are indexed or have type ANY. If a field is already
   indexed, then its type is already validated. If a field has
   type ANY, then any value is ok.

Closes #2800