Fix collation test on some platforms and disable collation in vinyl
It is appeared that ICU v50 supports CLRD-22 when ICU v51+ supports
CLRD-23+. One of the CLRD-23 changes comparing to CLRD-22 is
local script collates before other scripts. In general that means
that having different ICU version we get different order with
non-Latin locales. The main problem is that a run that was sorted
with one version of tarantool becomes unsorted in another.

Fix tests that depends on that order and disable collation in vinyl.