Cleanup engine lookup
To find an engine object, we always use engine_find(), which raises an
error if the engine was not found. Sometimes we know for sure that the
engine we are looking for exists, in which case we don't need to raise
an error. Let's introduce engine_by_name() that just looks up the engine
and does nothing else and use it when we are sure that the engine must
exist.