vinyl: complete dump if the index was dropped
If the index was dropped during dump, we abort the dump task so as not
to log information about zombie indexes to the vylog. However, we should
still notify the scheduler that its memory can be released by calling
vy_scheduler_complete_dump(). If we don't, lsregion won't be truncated
until another index is dumped. If there is no other vinyl indexes out
there or all other indexes have already been dumped, memory will never
be freed.

On debug builds, it will result in the following assertion failure:

  vy_scheduler.c:1319: vy_scheduler_peek_dump: Assertion `scheduler->dump_task_count > 0' failed.