vinyl: fix LSN assignment for indexes received during initial join
Since commit 7d67ec8aebd1 ("box: generate LSNs for rows received during
initial join"), we assign fake, monotonically growing LSNs to all
records received during initial join, because Vinyl requires all indexes
to have a unique LSN for identification in vylog. The problem is there's
a number of records (about 60) that have LSN 0 on the master - these are
bootstrap records. So if there are N such records, the index that has
LSN X on the master will have LSN (N + X) on the replica if sent during
initial join. But there may be another index with the same LSN on the
master (i.e. N + X). If this index is sent on final join or subscribe
stage, the replica will fail to make a checkpoint or recover, spitting
an error message similar to the one below:

  coio vy_log.c:2002 E> failed to process vylog record: create_index{index_lsn=68, space_id=12345, key_def=[0, 'unsigned'], }
  main/101/replica vy_log.c:1446 E> ER_INVALID_VYLOG_FILE: Invalid VYLOG file: Duplicate index id 68
  main/101/replica vy_log.c:2117 E> failed to load `./00000000000000000000.vylog'
  main/101/replica F> failed to create a checkpoint

Presently, the problem is 100% reproducible by

  test_run.py --conf vinyl engine/replica_join

To fix it, let's revert the aforementioned commit and instead assign
fake LSNs only to vinyl records, not to all records received during
initial join - after all it's vinyl that needs this, not memtx.

Actually, we already use the same technique for DML records in vinyl -
see vy_join_ctx::lsn - except there we assign fake LSNs on the master's
side. This was done for historical reasons - we could as well assign
them on the replica, but that would require some refactoring done later.
So let's use the same fake LSN counter for both cases.