box: allow to specify instance and replica set uuid
Add new box configuration options:

  box.cfg.instance_uuid
  box.cfg.replicaset_uuid

Both options take a uuid as a string. If set, they force tarantool to
set the specified UUIDs for instance and/or replica set on bootstrap.
On recovery, they are used to check UUIDs stored in the snapshot against
the ones specified by the user - on mismatch recovery is aborted.

Closes #2967