test: vinyl/low_quota.lua: pass size in args instead of using symlinks
Currently, to set the memory size in the low_quota.lua script, we use a
rather dirty trick: we create a symlink to low_quota.lua and code the
memory size in the name of the link (e.g. low_quota_2.lua). Instead we
can pass the memory size directly via the 'args' parameter so let's do
this.

While we are at it, let's also:

 - Fix a type in vinly/recovery_quota.test.lua:

     -- Disable dump and use all memory up to the limit.
     box.error.injection.set('ERRINJ_VY_RUN_WRITE', false)

   It should be 'true' obviously.

 - Remove useless switch to and from 'default' to restart a replica as
   the 'restart' command can be called from the replica script.