sql: add mechanism to emulate ROWID
It is worth mentioning, that there are two types of ephemeral tables:
one holds rows as they are inserted (all rows turn out to be distinct);
another adds ROWID column in order to hold equal rows.  Inasmuch as
original ROWID mechanism is to be removed, the purpose of this patch is
to emulate ROWID behavior for ephemeral tables.

 - Substituted all ephemeral table opcodes which don't rely on SQLite's
   ROWID with Tarantool ones.
 - Implemeted opcode OP_NextIdEphemeral to emulate behavior of ROWID
   feature.  In current implemetation ROWID is not hidden PK index, but
   just another one column to keep in table equal tuples.
 - Substituted old ephemeral tables with ROWID for INSERT INTO ...
   SELECT ... FROM statement. It utilizes ephemeral tables as
   intermediate holder for data from SELECT
   (before it is inserted to table).

Part of #2680