replication: do not wait for all masters on recovery
If one cluster node is down permanently for some reason, no other node
can restart - they will stall in box.cfg{} until all other nodes are up
and running. This complicates a tarantool cluster deployment in real
world scenarios.

To address this issue, let's complete the configuration as soon as
connections have been established connections to the number of hosts
specified by the new configuration option, box.cfg.replication_quorum,
assuming the rest will connect asynchronously. If the option is unset,
it defaults to the number of entries in box.cfg.replication so this
patch shouldn't affect the behavior of existing setups.

Closes #2958