call: fail requests that do not close transaction
Currently, if a CALL/EVAL request leaves an open transaction at return,
we silently rollback it and print a warning to the log mentioning the
function name or eval expression to facilitate further debugging. After
issue #946 was fixed, we can't do that anymore, because request input,
which stores CALL/EVAL parameters, may be discarded before request
completion and hence be unavailable for logging. Without additional
information pointing at the culprit, the log message is pointless (see
issue #1100). We could copy the arguments, but that would slow down CALL
execution, which can't be justified solely by the need of verbose
logging. So let's stop being lenient and fail requests that do not close
transaction at return. This should encourage negligent users to finally
fix their code.

Follow-up #946