Add replace/delete functions for ephemeral spaces
Ephemeral tables shouldn't be involved in any transaction routine,
since they are used only for internal needs. Moreover, ephemeral spaces
can be created and destroyed within one transaction, which may lead to
incorrect behaviour. Thus, ephemeral replace and delete don't take txn
as an argument.

Furthermore, delete operation should work on tuples with nulls in primary key.
As a rule, it is considered to be bad practice, due to ambiguity when
deleting tuple from space with several tuples containing nulls in PK.
On the other hand, ephemeral spaces are used only for internal needs,
so if it is guaranteed that no such situation occur, it is OK to allow
insertion nulls in PK. Hence, function which implements deletion on ephemeral
tables, doesn't check tuple on null fields.

Currently, supported only by memtx engine.