replication: reconnect applier on master rebootstrap
If one node of a cluster is rebootstrapped (i.e. restarted from an
empty directory with the same configuration), other replicas will
never try to reconnect to it - the appliers will simply stop with
the ER_REPLICASET_UUID_MISMATCH error. The only way to fix this is
reconfigure replication on all other nodes.

Let's fix this problem by reassigning an applier to a new replica
in case its UUID mismatches the UUID of the replica it is currently
assigned to.

Cannot write a test, because rebootstrap is unreliable - see #3108.

Closes #3112